requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2
django-allauth==65.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-htmx==1.23.0
django-tailwind==4.0.1
pillow==11.2.1
psycopg2-binary==2.9.10
qrcode==8.1
sqlparse==0.5.3
typing_extensions==4.13.2
gunicorn==22.0.0
whitenoise==6.6.0
